+++ src/main/java/domain/board/Board.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.board;
 
 import domain.piece.Piece;
 import domain.piece.PieceInfo;
