+++ src/main/java/domain/board/Country.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.board;
 
 public enum Country {
     HAN,
