+++ src/main/java/domain/piece/Chariot.java
@@ -1,6 +1,6 @@
 package domain.piece;
 
-import domain.Country;
+import domain.board.Country;
 
 public class Chariot extends StraightMovingPiece {
     public Chariot(Country country) {
