+++ src/main/java/domain/piece/General.java
@@ -1,6 +1,6 @@
 package domain.piece;
 
-import domain.Country;
+import domain.board.Country;
 
 public class General extends SingleMovingPiece {
     public General(Country country) {
