+++ src/main/java/domain/piece/PieceInfo.java
@@ -1,6 +1,6 @@
 package domain.piece;
 
-import domain.Country;
+import domain.board.Country;
 
 public record PieceInfo(
         PieceType pieceType,
